Fix clang warning

+++ tests/test-mstring.c
@@ -774,7 +774,9 @@ static void test0(void)
 
 static void test_rounding(void)
 {
-  const size_t max = -(m_str1ng_size_t)1;
+  // The final cast to size_t is due to clang which now gives
+  // warning even if the cast an integer to a bigger size(!)
+  const size_t max = (size_t) -(m_str1ng_size_t)1;
   for(size_t a = max/2; a <= max; a++) {
     unsigned char m, e;
     size_t x = (size_t) a + a/2;
